Project 1
1. Write a complete C++ program that holds information for coffee orders. The
program will define a struct for an order. The struct will contain a string for the
coffee name (Columbian, French Roast, Italian Blend, or House), a float for the
cost, an integer for the strength (1-light, 2-medium, 3-dark), an int for the amount
of sugar (0-none, 1, 2, 3, 4), a bool for cream or not cream. The program should
declare an instance of the struct, Next, ask the user to select coffee name, and then
all of the other information. All of the information will be stored in the
appropriate attributes of the struct. After the user enters all of the information for
an order The program should print the information to the screen in a neat tabular
form. The program should continue looping taking in information and printing
out information until the user chooses to quit.

----------------------------------------------

Modify the struct from lab 4 so that the values for coffee name are enumerated as 1=light,
2=medium, 3=dark. Also, add to the struct an attribute for coffee size that will be enumerated as
small=0, regular=1, medium=2, large=3, extra-large =4. The cost is calculated as follows small
$1.50, , regular $2.10, medium $2.50, large $3, extra-large $4.25. Add .25 if they choose French
roast or Italian blend, add .13 for each level of sugar, and add .08 for cream.
Create another struct for customer. The costumer has attributes for first name, order number
which is a value between 1 and 100 inclusive, and a coffee struct.
In the main program declare an array of customers that is size 100. Next create a main menu.
The menu will first ask if the customer wants a cup of coffee and provide choices for yes, no or
quit. If the customer chooses yes, the program should ask for the first name and store it in the
appropriate location. Next, the program should display the options and allow the user to enter
their preferences. Based on the customer’s input the cost should be calculated. Then the
customer’s order number (array index + 1 is good for this) should be displayed. The program
should keep each order in a separate array element. If in the main menu the customer chooses
“no” then the system should ask if they want to see their order. If they choose yes then the
program should ask for the order number and then display all of the appropriate information. The
program should loop until the user chooses to quit in the main menu.